backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# PDF processing
PyMuPDF==1.23.8
pdf2image==1.16.3
PyPDF2==3.0.1
pikepdf==10.0.2

# OCR - PaddleOCR for high accuracy
paddleocr>=2.7.0

# Image processing
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
requests==2.31.0

